#2642 Clean code
#2642 Flatrate Import: Support explicit enddate and terminated contracts
