+++ de.metas.contracts/src/main/java/de/metas/contracts/flatrate/impexp/FlatrateTermImportProcess.java
@@ -1,40 +1,23 @@
 package de.metas.contracts.flatrate.impexp;
 
-import java.math.BigDecimal;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.sql.Timestamp;
-import java.util.List;
 import java.util.Properties;
 
 import org.adempiere.ad.persistence.TableModelLoader;
 import org.adempiere.ad.trx.api.ITrx;
-import org.adempiere.bpartner.service.IBPartnerDAO;
-import org.adempiere.exceptions.AdempiereException;
 import org.adempiere.impexp.AbstractImportProcess;
 import org.adempiere.model.InterfaceWrapperHelper;
-import org.adempiere.model.PlainContextAware;
-import org.adempiere.util.Services;
 import org.adempiere.util.lang.IMutable;
-import org.adempiere.util.time.SystemTime;
-import org.compiere.model.I_AD_User;
 import org.compiere.model.I_C_BPartner;
-import org.compiere.model.I_C_UOM;
 import org.compiere.model.I_M_Product;
 import org.compiere.model.PO;
 import org.compiere.util.DB;
 
-import com.google.common.collect.Ordering;
-
-import de.metas.adempiere.model.I_C_BPartner_Location;
-import de.metas.contracts.IFlatrateBL;
 import de.metas.contracts.model.I_C_Flatrate_Conditions;
 import de.metas.contracts.model.I_C_Flatrate_Term;
 import de.metas.contracts.model.I_I_Flatrate_Term;
-import de.metas.contracts.model.X_C_Flatrate_Term;
 import de.metas.contracts.model.X_I_Flatrate_Term;
-import de.metas.product.IProductBL;
-import lombok.NonNull;
 
 /*
  * #%L
@@ -60,8 +43,12 @@ import lombok.NonNull;
 
 public class FlatrateTermImportProcess extends AbstractImportProcess<I_I_Flatrate_Term>
 {
-	private final transient IFlatrateBL flatrateBL = Services.get(IFlatrateBL.class);
-	private final transient IBPartnerDAO bpartnerDAO = Services.get(IBPartnerDAO.class);
+	private final FlatrateTermImportHelper flatRateImporter;
+
+	public FlatrateTermImportProcess()
+	{
+		flatRateImporter = FlatrateTermImportHelper.newInstance().setProcess(this);
+	}
 
 	@Override
 	public Class<I_I_Flatrate_Term> getImportModelClass()
@@ -90,102 +77,104 @@ public class FlatrateTermImportProcess extends AbstractImportProcess<I_I_Flatrat
 	@Override
 	protected void updateAndValidateImportRecords()
 	{
-		final String trxName = ITrx.TRXNAME_ThreadInherited;
 		final String sqlImportWhereClause = COLUMNNAME_I_IsImported + "<>" + DB.TO_BOOLEAN(true)
 				+ "\n " + getWhereClause();
 
-		//
-		// Update C_BPartner_ID
-		{
-			final String sqlSelectByValue = "select MIN(bp." + I_C_BPartner.COLUMNNAME_C_BPartner_ID + ")"
-					+ " from " + I_C_BPartner.Table_Name + " bp "
-					+ " where bp." + I_C_BPartner.COLUMNNAME_Value + "=i." + I_I_Flatrate_Term.COLUMNNAME_BPartnerValue
-					+ " and bp." + I_C_BPartner.COLUMNNAME_AD_Client_ID + "=i." + I_I_Flatrate_Term.COLUMNNAME_AD_Client_ID;
-			final String sql = "UPDATE " + I_I_Flatrate_Term.Table_Name + " i "
-					+ "\n SET " + I_I_Flatrate_Term.COLUMNNAME_C_BPartner_ID + "=(" + sqlSelectByValue + ")"
-					+ "\n WHERE " + sqlImportWhereClause
-					+ "\n AND i." + I_I_Flatrate_Term.COLUMNNAME_C_BPartner_ID + " IS NULL";
+		dbUpdateBPartnerIds(sqlImportWhereClause);
+		dbUpdateDropshipPartnerIds(sqlImportWhereClause);
+		dbUpdateC_Flatrate_Conditions_IDs(sqlImportWhereClause);
+		dbUpdateProductIdsByValue(sqlImportWhereClause);
+		dbUpdateProductIdsByName(sqlImportWhereClause);
+	}
+	
+	private void dbUpdateBPartnerIds(final String sqlImportWhereClause)
+	{
+		final String sqlSelectByValue = "select MIN(bp." + I_C_BPartner.COLUMNNAME_C_BPartner_ID + ")"
+				+ " from " + I_C_BPartner.Table_Name + " bp "
+				+ " where bp." + I_C_BPartner.COLUMNNAME_Value + "=i." + I_I_Flatrate_Term.COLUMNNAME_BPartnerValue
+				+ " and bp." + I_C_BPartner.COLUMNNAME_AD_Client_ID + "=i." + I_I_Flatrate_Term.COLUMNNAME_AD_Client_ID;
+		final String sql = "UPDATE " + I_I_Flatrate_Term.Table_Name + " i "
+				+ "\n SET " + I_I_Flatrate_Term.COLUMNNAME_C_BPartner_ID + "=(" + sqlSelectByValue + ")"
+				+ "\n WHERE " + sqlImportWhereClause
+				+ "\n AND i." + I_I_Flatrate_Term.COLUMNNAME_C_BPartner_ID + " IS NULL";
 
-			final int no = DB.executeUpdateEx(sql, trxName);
-			log.debug("Set C_BPartner_ID for {} records", no);
+		final int no = DB.executeUpdateEx(sql, ITrx.TRXNAME_ThreadInherited);
+		log.debug("Set C_BPartner_ID for {} records", no);
 
-			// Flag missing BPartners
-			markAsError("BPartner not found", I_I_Flatrate_Term.COLUMNNAME_C_BPartner_ID + " IS NULL"
-					+ "\n AND " + sqlImportWhereClause);
-		}
+		// Flag missing BPartners
+		markAsError("BPartner not found", I_I_Flatrate_Term.COLUMNNAME_C_BPartner_ID + " IS NULL"
+				+ "\n AND " + sqlImportWhereClause);
+	}
 
-		//
-		// Update DropShip_BPartner_ID
-		{
-			final String sqlSelectByValue = "select MIN(bp." + I_C_BPartner.COLUMNNAME_C_BPartner_ID + ")"
-					+ " from " + I_C_BPartner.Table_Name + " bp "
-					+ " where bp." + I_C_BPartner.COLUMNNAME_Value + "=i." + I_I_Flatrate_Term.COLUMNNAME_DropShip_BPartner_Value
-					+ " and bp." + I_C_BPartner.COLUMNNAME_AD_Client_ID + "=i." + I_I_Flatrate_Term.COLUMNNAME_AD_Client_ID;
-			final String sql = "UPDATE " + I_I_Flatrate_Term.Table_Name + " i "
-					+ "\n SET " + I_I_Flatrate_Term.COLUMNNAME_DropShip_BPartner_ID + "=(" + sqlSelectByValue + ")"
-					+ "\n WHERE " + sqlImportWhereClause
-					+ "\n AND i." + I_I_Flatrate_Term.COLUMNNAME_DropShip_BPartner_ID + " IS NULL";
+	private void dbUpdateDropshipPartnerIds(final String sqlImportWhereClause)
+	{
+		final String sqlSelectByValue = "select MIN(bp." + I_C_BPartner.COLUMNNAME_C_BPartner_ID + ")"
+				+ " from " + I_C_BPartner.Table_Name + " bp "
+				+ " where bp." + I_C_BPartner.COLUMNNAME_Value + "=i." + I_I_Flatrate_Term.COLUMNNAME_DropShip_BPartner_Value
+				+ " and bp." + I_C_BPartner.COLUMNNAME_AD_Client_ID + "=i." + I_I_Flatrate_Term.COLUMNNAME_AD_Client_ID;
+		final String sql = "UPDATE " + I_I_Flatrate_Term.Table_Name + " i "
+				+ "\n SET " + I_I_Flatrate_Term.COLUMNNAME_DropShip_BPartner_ID + "=(" + sqlSelectByValue + ")"
+				+ "\n WHERE " + sqlImportWhereClause
+				+ "\n AND i." + I_I_Flatrate_Term.COLUMNNAME_DropShip_BPartner_ID + " IS NULL";
 
-			final int no = DB.executeUpdateEx(sql, trxName);
-			log.debug("Set DropShip_BPartner_ID for {} records", no);
+		final int no = DB.executeUpdateEx(sql, ITrx.TRXNAME_ThreadInherited);
+		log.debug("Set DropShip_BPartner_ID for {} records", no);
 
-			// Flag missing DropShip BPartners
-			markAsError("DropShip BPartner not found",
-					I_I_Flatrate_Term.COLUMNNAME_DropShip_BPartner_ID + " IS NULL"
-							+ "\n AND " + I_I_Flatrate_Term.COLUMNNAME_DropShip_BPartner_Value + " IS NOT NULL" // only where the we have a key to match
-							+ "\n AND " + sqlImportWhereClause);
-		}
+		// Flag missing DropShip BPartners
+		markAsError("DropShip BPartner not found",
+				I_I_Flatrate_Term.COLUMNNAME_DropShip_BPartner_ID + " IS NULL"
+						+ "\n AND " + I_I_Flatrate_Term.COLUMNNAME_DropShip_BPartner_Value + " IS NOT NULL" // only where the we have a key to match
+						+ "\n AND " + sqlImportWhereClause);
+	}
 
-		//
-		// Update C_Flatrate_Conditions_ID
-		{
-			final String sqlSelectByValue = "select MIN(fc." + I_C_Flatrate_Conditions.COLUMNNAME_C_Flatrate_Conditions_ID + ")"
-					+ " from " + I_C_Flatrate_Conditions.Table_Name + " fc "
-					+ " where fc." + I_C_Flatrate_Conditions.COLUMNNAME_Name + "=i." + I_I_Flatrate_Term.COLUMNNAME_C_Flatrate_Conditions_Value
-					+ " and fc." + I_C_Flatrate_Conditions.COLUMNNAME_AD_Client_ID + "=i." + I_I_Flatrate_Term.COLUMNNAME_AD_Client_ID;
-			final String sql = "UPDATE " + I_I_Flatrate_Term.Table_Name + " i "
-					+ "\n SET " + I_I_Flatrate_Term.COLUMNNAME_C_Flatrate_Conditions_ID + "=(" + sqlSelectByValue + ")"
-					+ "\n WHERE " + sqlImportWhereClause
-					+ "\n AND i." + I_I_Flatrate_Term.COLUMNNAME_C_Flatrate_Conditions_ID + " IS NULL";
+	private void dbUpdateC_Flatrate_Conditions_IDs(final String sqlImportWhereClause)
+	{
+		final String sqlSelectByValue = "select MIN(fc." + I_C_Flatrate_Conditions.COLUMNNAME_C_Flatrate_Conditions_ID + ")"
+				+ " from " + I_C_Flatrate_Conditions.Table_Name + " fc "
+				+ " where fc." + I_C_Flatrate_Conditions.COLUMNNAME_Name + "=i." + I_I_Flatrate_Term.COLUMNNAME_C_Flatrate_Conditions_Value
+				+ " and fc." + I_C_Flatrate_Conditions.COLUMNNAME_AD_Client_ID + "=i." + I_I_Flatrate_Term.COLUMNNAME_AD_Client_ID;
+		final String sql = "UPDATE " + I_I_Flatrate_Term.Table_Name + " i "
+				+ "\n SET " + I_I_Flatrate_Term.COLUMNNAME_C_Flatrate_Conditions_ID + "=(" + sqlSelectByValue + ")"
+				+ "\n WHERE " + sqlImportWhereClause
+				+ "\n AND i." + I_I_Flatrate_Term.COLUMNNAME_C_Flatrate_Conditions_ID + " IS NULL";
 
-			final int no = DB.executeUpdateEx(sql, trxName);
-			log.debug("Set C_Flatrate_Conditions_ID for {} records", no);
+		final int no = DB.executeUpdateEx(sql, ITrx.TRXNAME_ThreadInherited);
+		log.debug("Set C_Flatrate_Conditions_ID for {} records", no);
 
-			// Flag missing flatrate conditions
-			markAsError("Flatrate conditions not found", I_I_Flatrate_Term.COLUMNNAME_C_Flatrate_Conditions_ID + " IS NULL"
-					+ "\n AND " + sqlImportWhereClause);
-		}
+		// Flag missing flatrate conditions
+		markAsError("Flatrate conditions not found", I_I_Flatrate_Term.COLUMNNAME_C_Flatrate_Conditions_ID + " IS NULL"
+				+ "\n AND " + sqlImportWhereClause);
+	}
 
-		//
-		// Update M_Product_ID by Value
-		{
-			final String sqlSelectByValue = "select MIN(bp." + I_M_Product.COLUMNNAME_M_Product_ID + ")"
-					+ " from " + I_M_Product.Table_Name + " bp "
-					+ " where bp." + I_M_Product.COLUMNNAME_Value + "=i." + I_I_Flatrate_Term.COLUMNNAME_ProductValue
-					+ " and bp." + I_M_Product.COLUMNNAME_AD_Client_ID + "=i." + I_I_Flatrate_Term.COLUMNNAME_AD_Client_ID;
-			final String sql = "UPDATE " + I_I_Flatrate_Term.Table_Name + " i "
-					+ "\n SET " + I_I_Flatrate_Term.COLUMNNAME_M_Product_ID + "=(" + sqlSelectByValue + ")"
-					+ "\n WHERE " + sqlImportWhereClause
-					+ "\n AND i." + I_I_Flatrate_Term.COLUMNNAME_M_Product_ID + " IS NULL";
+	private void dbUpdateProductIdsByValue(final String sqlImportWhereClause)
+	{
+		final String sqlSelectByValue = "select MIN(bp." + I_M_Product.COLUMNNAME_M_Product_ID + ")"
+				+ " from " + I_M_Product.Table_Name + " bp "
+				+ " where bp." + I_M_Product.COLUMNNAME_Value + "=i." + I_I_Flatrate_Term.COLUMNNAME_ProductValue
+				+ " and bp." + I_M_Product.COLUMNNAME_AD_Client_ID + "=i." + I_I_Flatrate_Term.COLUMNNAME_AD_Client_ID;
+		final String sql = "UPDATE " + I_I_Flatrate_Term.Table_Name + " i "
+				+ "\n SET " + I_I_Flatrate_Term.COLUMNNAME_M_Product_ID + "=(" + sqlSelectByValue + ")"
+				+ "\n WHERE " + sqlImportWhereClause
+				+ "\n AND i." + I_I_Flatrate_Term.COLUMNNAME_M_Product_ID + " IS NULL";
 
-			final int no = DB.executeUpdateEx(sql, trxName);
-			log.debug("Set M_Product_ID for {} records (by Value)", no);
-		}
-		// Update M_Product_ID by Name
-		{
-			final String sqlSelectByValue = "select MIN(bp." + I_M_Product.COLUMNNAME_M_Product_ID + ")"
-					+ " from " + I_M_Product.Table_Name + " bp "
-					+ " where bp." + I_M_Product.COLUMNNAME_Name + "=i." + I_I_Flatrate_Term.COLUMNNAME_ProductValue
-					+ " and bp." + I_M_Product.COLUMNNAME_AD_Client_ID + "=i." + I_I_Flatrate_Term.COLUMNNAME_AD_Client_ID;
-			final String sql = "UPDATE " + I_I_Flatrate_Term.Table_Name + " i "
-					+ "\n SET " + I_I_Flatrate_Term.COLUMNNAME_M_Product_ID + "=(" + sqlSelectByValue + ")"
-					+ "\n WHERE " + sqlImportWhereClause
-					+ "\n AND i." + I_I_Flatrate_Term.COLUMNNAME_M_Product_ID + " IS NULL";
+		final int no = DB.executeUpdateEx(sql, ITrx.TRXNAME_ThreadInherited);
+		log.debug("Set M_Product_ID for {} records (by Value)", no);
+	}
 
-			final int no = DB.executeUpdateEx(sql, trxName);
-			log.debug("Set M_Product_ID for {} records (by Name)", no);
+	private void dbUpdateProductIdsByName(final String sqlImportWhereClause)
+	{
+		final String sqlSelectByValue = "select MIN(bp." + I_M_Product.COLUMNNAME_M_Product_ID + ")"
+				+ " from " + I_M_Product.Table_Name + " bp "
+				+ " where bp." + I_M_Product.COLUMNNAME_Name + "=i." + I_I_Flatrate_Term.COLUMNNAME_ProductValue
+				+ " and bp." + I_M_Product.COLUMNNAME_AD_Client_ID + "=i." + I_I_Flatrate_Term.COLUMNNAME_AD_Client_ID;
+		final String sql = "UPDATE " + I_I_Flatrate_Term.Table_Name + " i "
+				+ "\n SET " + I_I_Flatrate_Term.COLUMNNAME_M_Product_ID + "=(" + sqlSelectByValue + ")"
+				+ "\n WHERE " + sqlImportWhereClause
+				+ "\n AND i." + I_I_Flatrate_Term.COLUMNNAME_M_Product_ID + " IS NULL";
 
-		}
+		final int no = DB.executeUpdateEx(sql, ITrx.TRXNAME_ThreadInherited);
+		log.debug("Set M_Product_ID for {} records (by Name)", no);
+
 		// Flag missing product
 		markAsError("Product not found", I_I_Flatrate_Term.COLUMNNAME_M_Product_ID + " IS NULL"
 				+ "\n AND " + sqlImportWhereClause);
@@ -211,127 +200,8 @@ public class FlatrateTermImportProcess extends AbstractImportProcess<I_I_Flatrat
 	@Override
 	protected ImportRecordResult importRecord(final IMutable<Object> state, final I_I_Flatrate_Term importRecord) throws Exception
 	{
-		final I_M_Product product = importRecord.getM_Product();
-
-		final I_C_Flatrate_Term contract = flatrateBL.createTerm(
-				PlainContextAware.newWithThreadInheritedTrx(), // context
-				importRecord.getC_BPartner(), // bpartner
-				importRecord.getC_Flatrate_Conditions(), // conditions
-				importRecord.getStartDate(), // startDate
-				(I_AD_User)null, // userInCharge
-				product, // product
-				false // completeIt
-		);
-		if (contract == null)
-		{
-			throw new AdempiereException("contract not created");
-		}
-
-		setDropShipBPartner(importRecord, contract);
-		setDropShipLocation(contract);
-		contract.setM_Product(product);
-		setUOM(contract, product);
-		contract.setPriceActual(importRecord.getPrice());
-		setPlannedQtyPerUnit(importRecord, contract);
-		setEndDate(importRecord, contract);
-		setMasterStartdDate(importRecord, contract);
-		setMasterEnddDate(importRecord, contract);
-		// important to ended if needed, before saving
-		endContractIfNeeded(importRecord, contract);
-		InterfaceWrapperHelper.save(contract);
-		//
-		// Complete the subscription/contract
-		if (!isEndedContract(importRecord))
-		{
-			flatrateBL.complete(contract);
-		}
-		//
-		// Link back the contract to current import record
-		importRecord.setC_Flatrate_Term(contract);
-		//
+		flatRateImporter.importRecord(importRecord);
 		return ImportRecordResult.Inserted;
-	}
-
-	private void setDropShipBPartner(@NonNull final I_I_Flatrate_Term importRecord, @NonNull final I_C_Flatrate_Term contract)
-	{
-		final int dropShipBPartnerId = importRecord.getDropShip_BPartner_ID() > 0 ? importRecord.getDropShip_BPartner_ID() : importRecord.getC_BPartner_ID();
-		if (dropShipBPartnerId <= 0)
-		{
-			throw new AdempiereException("DropShip BPartner not found");
-		}
-		contract.setDropShip_BPartner_ID(dropShipBPartnerId);
-	}
-
-	private void setDropShipLocation(@NonNull final I_C_Flatrate_Term contract)
-	{
-		int dropShipBPartnerId = contract.getDropShip_BPartner_ID();
-		final I_C_BPartner_Location dropShipBPLocation = findBPartnerShipToLocation(dropShipBPartnerId);
-		if (dropShipBPLocation != null)
-		{
-			contract.setDropShip_Location(dropShipBPLocation);
-		}
-	}
-
-	private void setUOM(@NonNull final I_C_Flatrate_Term contract, @NonNull final I_M_Product product)
-	{
-		final I_C_UOM uom = Services.get(IProductBL.class).getStockingUOM(product);
-		contract.setC_UOM(uom);
-	}
-
-	private void setPlannedQtyPerUnit(@NonNull final I_I_Flatrate_Term importRecord, @NonNull final I_C_Flatrate_Term contract)
-	{
-		if (importRecord.getQty() != null && importRecord.getQty().intValue() > 0)
-		{
-			contract.setPlannedQtyPerUnit(importRecord.getQty());
-		}
-	}
-
-	private void setEndDate(@NonNull final I_I_Flatrate_Term importRecord, @NonNull final I_C_Flatrate_Term contract)
-	{
-		if (importRecord.getEndDate() != null)
-		{
-			contract.setEndDate(importRecord.getEndDate());
-		}
-	}
-
-	private void setMasterStartdDate(@NonNull final I_I_Flatrate_Term importRecord, @NonNull final I_C_Flatrate_Term contract)
-	{
-		if (importRecord.getMasterStartDate() != null)
-		{
-			contract.setMasterStartDate(importRecord.getMasterStartDate());
-		}
-	}
-
-	private void setMasterEnddDate(@NonNull final I_I_Flatrate_Term importRecord, @NonNull final I_C_Flatrate_Term contract)
-	{
-		if (importRecord.getMasterEndDate() != null)
-		{
-			contract.setMasterEndDate(importRecord.getMasterEndDate());
-		}
-	}
-
-	private boolean isEndedContract(@NonNull final I_I_Flatrate_Term importRecord)
-	{
-		final Timestamp contractEndDate = importRecord.getEndDate();
-		final Timestamp today = SystemTime.asDayTimestamp();
-		if (contractEndDate != null && today.after(contractEndDate))
-		{
-			return true;
-		}
-		return false;
-	}
-
-	private void endContractIfNeeded(@NonNull final I_I_Flatrate_Term importRecord, @NonNull final I_C_Flatrate_Term contract)
-	{
-		if (isEndedContract(importRecord))
-		{
-			contract.setContractStatus(X_C_Flatrate_Term.CONTRACTSTATUS_Quit);
-			contract.setNoticeDate(contract.getEndDate());
-			contract.setIsAutoRenew(false);
-			contract.setProcessed(true);
-			contract.setDocAction(X_C_Flatrate_Term.DOCACTION_None);
-			contract.setDocStatus(X_C_Flatrate_Term.DOCSTATUS_Completed);
-		}
 	}
 
 	@Override
@@ -344,33 +214,4 @@ public class FlatrateTermImportProcess extends AbstractImportProcess<I_I_Flatrat
 		importRecord.setProcessed(true);
 		InterfaceWrapperHelper.save(importRecord);
 	}
-
-	private I_C_BPartner_Location findBPartnerShipToLocation(final int bpartnerId)
-	{
-		final List<I_C_BPartner_Location> bpLocations = bpartnerDAO.retrieveBPartnerLocations(getCtx(), bpartnerId, ITrx.TRXNAME_None);
-		if (bpLocations.isEmpty())
-		{
-			return null;
-		}
-		else if (bpLocations.size() == 1)
-		{
-			return bpLocations.get(0);
-		}
-		else
-		{
-			final I_C_BPartner_Location bpLocation = bpLocations.stream()
-					.filter(I_C_BPartner_Location::isShipTo)
-					.sorted(Ordering.natural().onResultOf(bpl -> bpl.isShipToDefault() ? 0 : 1))
-					.findFirst().get();
-			if (bpLocation.isShipToDefault())
-			{
-				return bpLocation;
-			}
-			else
-			{
-				return null;
-			}
-		}
-	}
-
 }
